Fix build

+++ import_from_betagouv.sh
@@ -7,16 +7,18 @@ function remove_old_files {
    pwd
    ls
    mv _index.md _index.md.bak
-   rm *.md
+   if [ find . -type f -name '*.md' ]; then
+      rm *.md
+   fi
    mv _index.md.bak _index.md
 }
 
 cd "$startup_dir"
 remove_old_files
 
-cp "$ROOT/artifacts/startups/." "$startup_dir"
+cp -R "$ROOT/artifacts/startups/" "$startup_dir/"
 
 cd "$people_dir"
 remove_old_files
 
-cp "$ROOT/artifacts/authors/." "$people_dir"
+cp -R "$ROOT/artifacts/authors/" "$people_dir/"
